Use dangerouslySetInnerHTML in ProblemDescription
This means we can specify markup instead of just text (which React
would escape).

The motivation is to be able to support the figure required in the
statement of 2004, question 3.

The large number of diffs to problems.json is because I had to
replace the < and > symbols with their latex codes \lt, \gt now
that these strings are written directly to the DOM.